chore(guards): keep three files under the size ratchet, and stop counting a required issue field as prose
The new event's docs pushed kind.rs, recall.rs and textline.rs past the
1500-line guard. Each is trimmed in place rather than baselined, except
textline's new steering_dropped, which gets textline/steering.rs beside
the gate and memory modules it matches.

check-prose counted `ConsumerPosture::RecordedOnly { issue: "#1234" }`
as an issue number in prose. The audit requires that value, so no new
event of that posture could be added without raising the file's count
with nothing a reader could delete. The exemption is the value form
alone; a number in a sentence still fails.

Also drops the bare issue number from the skills-section advisory, which
told the person reading it nothing, and updates the assertion that
pinned it.
