+++ crates/stella-tui/src/transcript_build.rs
@@ -502,9 +502,9 @@ fn note_kind(event: &AgentEvent) -> NoteKind {
         // colors both of these (`render/entry.rs`), so the plain and
         // exported forms must not mute them to the same glyph (`#5748`).
         AgentEvent::Error { .. } | AgentEvent::SteeringWithheld { .. } => NoteKind::Alert,
-        // Alert too, and for the same reason: the live deck draws it in
-        // `theme::WARNING`, so muting it here would make the exported
-        // transcript disagree with the screen it is a record of.
+        // Alert too, and for the same reason. The live deck draws it in
+        // `theme::WARNING`. Muting it here would make the exported
+        // transcript disagree with the screen it records.
         AgentEvent::SteeringDropped { .. } => NoteKind::Alert,
         // `Steered` stays in the wildcard on purpose: the live deck folds it
         // into a full user-turn row, not a note, and the plain `Note` model
