+++ docs/reference/diagnostics.md
@@ -345,6 +345,23 @@ The run entered a named `stage`. These records partition the timeline; the gap b
 
 The human steered the run mid-turn. The steering text is content and stays off this plane; the record marks where in the timeline the run's direction changed — behaviour before and after it should not be compared as one run.
 
+### `agent.steering.dropped`
+
+<!-- facts:begin — generated by `make diag-reference`; do not edit by hand -->
+- **Level:** `debug`
+- **Emitted from:** `crates/stella-cli/src/diag_bridge.rs`
+- **Fields:** `seq`
+<!-- facts:end -->
+
+A steering candidate did not fit its budget, so it never reached the prompt.
+One record per refusal. When a run keeps ignoring a workspace skill, count
+these rather than guess. A turn that refuses several every time is a turn
+whose budgets are set too low for what the workspace declares.
+
+The handle is text the workspace wrote, so it stays off this plane. The
+transcript's `⚠ steering` rows name the candidate and the knob that widens
+the budget.
+
 ### `agent.steering.withheld`
 
 <!-- facts:begin — generated by `make diag-reference`; do not edit by hand -->
